+++ src/helpers/errorHandler.ts
@@ -3,7 +3,6 @@ import log from 'loglevel'
 import { Bus } from '@/helpers'
 import i18n from '@/localization'
 
-// FIXME
 export class ErrorHandler {
   static process(error: Error | unknown, errorMessage = ''): void {
     const msgTranslation = errorMessage || ErrorHandler._getErrorMessage(error)